+++ src/blacken_docs/__init__.py
@@ -4,6 +4,7 @@ import argparse
 import contextlib
 import re
 import textwrap
+from bisect import bisect
 from typing import Generator
 from typing import Match
 from typing import Sequence
@@ -87,6 +88,16 @@ PYTHONTEX_RE = re.compile(
 )
 INDENT_RE = re.compile("^ +(?=[^ ])", re.MULTILINE)
 TRAILING_NL_RE = re.compile(r"\n+\Z", re.MULTILINE)
+ON_OFF = r"blacken-docs:(on|off)"
+ON_OFF_COMMENT_RE = re.compile(
+    # Markdown
+    rf"(?:^\s*<!-- {ON_OFF} -->$)|"
+    # rST
+    rf"(?:^\s*\.\. +{ON_OFF}$)|"
+    # LaTeX
+    rf"(?:^\s*% {ON_OFF}$)",
+    re.MULTILINE,
+)
 
 
 class CodeBlockError:
@@ -103,6 +114,29 @@ def format_str(
 ) -> tuple[str, Sequence[CodeBlockError]]:
     errors: list[CodeBlockError] = []
 
+    off_ranges = []
+    off_start = None
+    for comment in re.finditer(ON_OFF_COMMENT_RE, src):
+        # Check for the "off" value across the multiple (on|off) groups.
+        if "off" in comment.groups():
+            if off_start is None:
+                off_start = comment.start()
+        else:
+            if off_start is not None:
+                off_ranges.append((off_start, comment.end()))
+                off_start = None
+    if off_start is not None:
+        off_ranges.append((off_start, len(src)))
+
+    def _within_off_range(code_range: tuple[int, int]) -> bool:
+        index = bisect(off_ranges, code_range)
+        try:
+            off_start, off_end = off_ranges[index - 1]
+        except IndexError:
+            return False
+        code_start, code_end = code_range
+        return code_start >= off_start and code_end <= off_end
+
     @contextlib.contextmanager
     def _collect_error(match: Match[str]) -> Generator[None, None, None]:
         try:
@@ -111,6 +145,8 @@ def format_str(
             errors.append(CodeBlockError(match.start(), e))
 
     def _md_match(match: Match[str]) -> str:
+        if _within_off_range(match.span()):
+            return match[0]
         code = textwrap.dedent(match["code"])
         with _collect_error(match):
             code = black.format_str(code, mode=black_mode)
@@ -118,6 +154,8 @@ def format_str(
         return f'{match["before"]}{code}{match["after"]}'
 
     def _rst_match(match: Match[str]) -> str:
+        if _within_off_range(match.span()):
+            return match[0]
         lang = match["lang"]
         if lang is not None and lang not in PYGMENTS_PY_LANGS:
             return match[0]
@@ -132,6 +170,8 @@ def format_str(
         return f'{match["before"]}{code.rstrip()}{trailing_ws}'
 
     def _rst_literal_blocks_match(match: Match[str]) -> str:
+        if _within_off_range(match.span()):
+            return match[0]
         if not match["code"].strip():
             return match[0]
         min_indent = min(INDENT_RE.findall(match["code"]))
@@ -190,17 +230,23 @@ def format_str(
         return code
 
     def _md_pycon_match(match: Match[str]) -> str:
+        if _within_off_range(match.span()):
+            return match[0]
         code = _pycon_match(match)
         code = textwrap.indent(code, match["indent"])
         return f'{match["before"]}{code}{match["after"]}'
 
     def _rst_pycon_match(match: Match[str]) -> str:
+        if _within_off_range(match.span()):
+            return match[0]
         code = _pycon_match(match)
         min_indent = min(INDENT_RE.findall(match["code"]))
         code = textwrap.indent(code, min_indent)
         return f'{match["before"]}{code}'
 
     def _latex_match(match: Match[str]) -> str:
+        if _within_off_range(match.span()):
+            return match[0]
         code = textwrap.dedent(match["code"])
         with _collect_error(match):
             code = black.format_str(code, mode=black_mode)
@@ -208,6 +254,8 @@ def format_str(
         return f'{match["before"]}{code}{match["after"]}'
 
     def _latex_pycon_match(match: Match[str]) -> str:
+        if _within_off_range(match.span()):
+            return match[0]
         code = _pycon_match(match)
         code = textwrap.indent(code, match["indent"])
         return f'{match["before"]}{code}{match["after"]}'